requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.4
django-braces==1.13.0
django-cors-headers==2.4.0
django-import-export==1.2.0
django-mockups==0.4.8
django-oauth-toolkit==1.2.0
django-phonenumber-field==2.1.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-storages==1.7.1
djangorestframework==3.9.0
dropbox==9.3.0
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==1.0.1
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==2.1.0
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.5.12
phonenumbers==8.10.2
Pillow==5.4.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
PySocks==1.6.8
python-dateutil==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
tablib==0.12.1
text-unidecode==1.2
twilio==6.23.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
